pyproject.toml

[tool.poetry]
name = "nql-on-aws"
package-mode = false
version = "0.1.0"
description = "Querying database with Natural Language"
authors = ["y"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.39.0"
sqlalchemy = "^2.0.35"
sentence-transformers = "^3.1.1"
pyyaml = "^6.0.2"
psycopg2-binary = "^2.9.9"
openai = "^1.51.0"
langchain-huggingface = "^0.1.0"
langchain-experimental = "^0.3.2"
langchain-openai = "^0.2.1"
langchain-community = "^0.3.1"
langchain = "^0.3.1"
chromadb = "^0.5.11"
botocore = "^1.35.31"
boto3 = "^1.35.31"
awscli = "^1.34.31"
protobuf = "<=3.20.1"
kaleido = {version = "0.2.1", optional = true}
vanna = {extras = ["chromadb", "openai", "postgres"], version = "^0.7.3"}
chainlit = "^1.2.0"

[tool.poetry.group.dev.dependencies]
# Debugging
ipdb = "^0.13.9"
ipython = "^8.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
